ref(trimSlug): Move into own file & add test
